requirements.txt
pymongo
motor
requests
dnspython
tgcrypto
asyncio
python-dotenv
aiohttp
cloudscraper
git+https://github.com/Mayuri-Chan/pyrofork
google-generativeai
